requirements.txt
Django==1.11.1
djangorestframework==3.6.2
psycopg2==2.7.1
requests==2.13.0
